% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.R
\name{bn_avg_price}
\alias{bn_avg_price}
\title{Average price (5-min VWAP)}
\usage{
bn_avg_price(symbol, base = "us")
}
\arguments{
\item{symbol}{Trading pair}

\item{base}{"us" or "global"}
}
\value{
tibble: symbol, mins, price
}
\description{
Average price (5-min VWAP)
}
