% Generated by build script
\name{bn_trades}
\alias{bn_trades}
\title{Recent trades}
\usage{
bn_trades()
}
\arguments{
\item{symbol}{Trading pair}
\item{limit}{Number of trades (default 100, max 1000)}
\item{base}{"us" or "global"}
}
\value{tibble: id, price, qty, quote_qty, time, is_buyer_maker, symbol}
\description{Recent trades}
