% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bn.R
\name{bn_price}
\alias{bn_price}
\title{Current price for one or all symbols}
\usage{
bn_price(symbol = NULL, base = "us")
}
\arguments{
\item{symbol}{Trading pair (e.g. "BTCUSD", "ETHUSD"). NULL = all symbols.}

\item{base}{API region: "us" (default, Binance.US) or "global" (Binance.com)}
}
\value{
tibble: symbol, price
}
\description{
Current price for one or all symbols
}
